mononoke: inital support for backsyncing merges
Summary:
Now backsyncer should support syncing merges from a large repo to a small repo.

For more details see https://fb.quip.com/s8fYAOxEohtJ, but TLDR is that:
1) Syncing of merge commits from small repo to a large repo is not supported, at least for now.
Because of that if anyone wants to make a merge then it should be done in
megarepo directly

2) At the moment "naive" commit rewriting is used i.e. bonsai file changes are are rewritten directly. However that shouldn't be a difference in only few edge cases, and for all other cases naive rewriting should work for both merge and non-merge commits. [1]

3) Current implementation always syncs a merge commit to all small repos even if **merge commit is empty in this small repo**. This is mainly done for simplicity sake, we can change it later if necessary.

[1] Note that we still want to implement corrent manifest-based commit
rewriting, but for now we can start with the easy case.

Reviewed By: farnz

Differential Revision: D18743421

fbshipit-source-id: 4948e521192eb66dfe2171c59c55232ca93114eb