eden: redirect: use `hdiutil detach` to unmount images
Summary:
unmount will unmount from the VFS but still leave the image
associated with the disk image machinery and visible to `hdiutil info`.

This leaves resources associated with the mount even after it has
logically gone away, and that is undesirable.

This diff switches to using `detach` instead which unmounts and removes
the image from the disk image state.

The effect of this is that running `hg co null ; eden redirect fixup`
will completely unmount the redirections.

However, it doesn't cause them to be unmounted when the eden daemon is
stopped: that is something to address in a separate diff.

Also worth noting: only HFS+ mounts get detached successfully using this method.  APFS mounts will remain regardless, so its another reason for switching to HFS+ in D18795799.

Reviewed By: pkaush

Differential Revision: D18795800

fbshipit-source-id: dfc86d86016a0c78e67f6ae2651db681669a5b14