eden: call `eden redirect unmount` at unmount time
Summary:
The intent is to allow macOS to make a best effort attempt
to detach any disk images associated with redirections.

D18795800 taught redirect to detach rather than a simple unmount,
and we need to give it an opportunity to detach when we unmount
gracefully so that all resources are released on shutdown.

The implementation of this diff ties into the same mechanism that
we use to shutdown the buck daemon when we unmount or restart which
has the nice side effect of showing the output from the underlying
`eden redirect unmount` command.

In an earlier iteration of this diff I tried making the server code
run `eden redirect unmount` but it happened too late in the process
to be effective: the thrift server had already been torn down in the
shutdown case and we need privs to perform the unmount.

Reviewed By: pkaush

Differential Revision: D18804080

fbshipit-source-id: 0b409130752121c56a46c9b2e46b50e5abee8200