manifest: fix gettreepack for commit changing file to folder
Summary:
When changing a file to a folder in a commit this function would try to query
the storage using the path and the id of the file. Storage would not have
such an entry and would cause the operation to fail.

Fixing the issue by filtering out ids of files in the section for known
manifests. Filtering already occurs in the section for transmitted manifest.

Reviewed By: quark-zju

Differential Revision: D18797368

fbshipit-source-id: 2ed727b5eaaab2f17cd988d6c8a3a1055ac8bd2a