mononoke: add docstrings for rewrite_commit and sync_commit
Summary:
I always find it confusing, which one is responsible for what, so let's
add docstrings, which are (hopefully :P) easier to read than code.

I also renamed `rewrite_path` into `mover`, as that seems to be the accepted
name in other places and matches the type name. I am fine removing this rename.

Reviewed By: farnz

Differential Revision: D18792593

fbshipit-source-id: 35e623dfa750a28133d94bf2875cbd860c6a54ac