mononoke/manifest: introduce `Manifest` based `bonsai_diff`
Summary:
This adds an implementation of `bonsai_diff` (i.e. go from manifest and parent
manifest to list of Bonsai file changes). based on the `Manifest` trait. The
goal of this change is to make it possible to use `bonsai_diff` on any
`Manifest`, and in particular on things that are not `Box<dyn HgEntry>`.

This change also introduces a slightly lower-level `Loadable`-like trait for
the `Manifest` crate: `StoreLoadable`. `StoreLoadable` is basically `Loadable`
relative to a given store. It includes a blanket implementation for all
`Loadable`s (which are `StoreLoadable` relative to a `Blobstore`). Later in
this stack, I've used this to allow loading from a git repository.

Notably, I'm planning to use this to blobimport git repositories (later in this
stack). I'm also planning to update callsites.

Reviewed By: StanislavGlebik

Differential Revision: D18683685

fbshipit-source-id: d61a733af30465746cca7672e325d7cf00d0b8bf