mononoke/manifest: introduce `gitimport`
Summary:
This adds a `gitimport` binary to import small git repositories into Mononoke.
This isn't intended as a production-grade product (notably, there is no Git <->
Bonsai mapping right now, and we import everything in one go, there are a few
other TODOs here and there), but it's a meaningful-enough starting point that I
think it's worth reviewing now.

I've used this to import a git repository into Mononoke, and I was able to
clone the repository through Mercurial afterwards. The contents of the git
repository and the derived Mercurial repository were identical

Differential Revision: D18683684

fbshipit-source-id: 7375824d21562f5ba6f2dbfee83f7c34ba8057be