Delete failure_ext::bail_err in favor of bail
Summary:
In the conversion of failure_ext from failure to anyhow, the failure_ext::bail_err macro was preserved in order to minimize the amount of code that needed to be migrated all at once. However, it's simply a re-export of anyhow::bail. This codemod deletes failure_ext::bail_err and replaces all references to it by failure_ext::bail.

Progression:

1. Originally, failure_ext::bail_err is a macro_rules macro defined in failure_ext.
2. In D18577341 we replaced failure_ext::bail_err with a re-export of anyhow::bail. At a high level: failure_ext serves as a shim to make the API of anyhow look more like the API of failure temporarily.
3. *(you are here)* We replace all uses of failure_ext::bail_err by failure_ext::bail. At a high level: make the API of failure_ext look more like the API of anyhow.
4. Coming soon: once failure_ext looks more like anyhow from this and other changes, we replace dependencies on failure_ext by direct dependencies on anyhow.

Reviewed By: Imxset21

Differential Revision: D18828417

fbshipit-source-id: b3de50e956345a8c600078d57f7ff0bb9ea90aa1