make `eden top` ignore errors when writing data to curses
Summary:
Add a wrapper function around `addnstr()` that silently ignores any errors.

When writing to the last cell of a terminal curses will successfully display
the data, but report an error since the logical cursor cannot be moved to the
next cell.  This error is not important and can be ignored.

The errors reported by the `curses` module unfortunately are not very useful
(they normally just report `addnwstr() returned ERR`).  For now we silently
ignore all errors, although this diff does contain some stub code that can be
enabled to examine them during development.

Reviewed By: wez

Differential Revision: D18826917

fbshipit-source-id: a4c5394b49aea1334def1d047eec7b5e743fb9f2