mononoke: hopefully remove enough git from the contbuild
Summary:
I'm going to temporarily back out parts of the git support diffs from master.

Long story short:

- Those diffs depend on libgit2-rs.
- libgit2-rs depends on PCRE, but it bundles it in its own build. This causes
  our mode/opt builds to crash when building a binary that depends on
  libgit2-rs and well as PCRE (the latter is a group that consists of "most
  things", so basically a libgit2-rs dependent binary will fail to build in
  mode/opt).

Now, we've seen this before — I had to fix the exact same problem with lz4
builds for example, and Simon did it with fuse. The fix is to make sure we
don't build ad-hoc binaries and bundle them in. So, I imported libgit 0.28 into
tp2.

Unfortunately, you can only have a single version of any given tp2 library in
fbcode, and it turns out pygit2 is dependent on a different version of libgit2
(0.27.5). This would normally be a simple matter of updating pygit2 ... but I
just discovered that we have effectively forked pygit2 internally �

Anyhow, at this point the next step is to figure out how to backport and test
pygit2 with libgit 0.28, but since I'm not going to be able to fix this super
quickly, I'll take git parts out until the contbuild can run again � I
submitted a diff earlier today to take it out of the contbuild, but I'll
probably have to take the integration tests out as well to get the build to run
again.

Reviewed By: ikostia

Differential Revision: D18855383

fbshipit-source-id: e7f54f4a27f41ee65abb291e43e7ff3c78cb4183