mononoke: add warmup functions for backfilling derived data
Summary:
Deriving data for a lot of commits is slow, mainly because in order to derive
data for a child you need to first derive data for all ancestors sequentially.
So if we need to derive data for a stack of 1K commits and deriving for a single commit
takes 1 second then in total it will take 15 mins.

In order to make derivation faster backfiller does a few things:
1) It derives data for CHUNK_SIZE commits at once
1) For all of those commits writes to blobstore are batched i.e. all writes
first go to memory the writes are sent to blobstore all at once.
2) It warms up the caches by reading all the necessary data for all CHUNK_SIZE
commits at once.

Previously we were only prefetching file content. However for some derived data (in particular,
for fastlog batches and for deleted file manfiest) warming up unodes makes a huge difference.

This diff adds exactly that.

Reviewed By: krallin

Differential Revision: D18834327

fbshipit-source-id: 1d68636a7f84bba93d20a5d559ed826cb9415ea2