mononoke: make backfill_derived_data safer
Summary:
backfill_derived_data had a footgun. It was writing all the blobs in parallel,
including blobs for derived data mapping. However these blobs should be written
to the blobstore only if all corresponding blobs were saved.

For example, when deriving unodes we deriving manifests entries, file entries
and mapping entries. Previously all of these blobs were written in parallel,
and so it was possible that file entry write failed but mapping entry
succeeded. That causes derived data to be corrupted because the invariant
"mapping entry is written only after all blobs for this commit and ancestors are
saved" is no longer holds.

This diff fixes it by ensuring the correct order of blobstore writes. Note
however that after this fix it's still possible that mapping entry for a child was written
while mapping entry for a parent didn't. But this shouldn't be a problem - in
the worst case it might trigger useless regeneration of derived data.

Reviewed By: krallin

Differential Revision: D18882252

fbshipit-source-id: 4a3cac3e7c5ee5cc890980014ae16aee167397fc