mononoke: use implicit delete detection in commit rewriting
Summary:
Implicit deletes are a challenge for commit rewriting when non-prefix-free commitsyncmaps are allowed.

Consider the following path rewriting config:
```
a/b => c
a => d
```
(all paths, which start with `a/b` should have this prefix replaced with `c`, while all other paths, which start with just `a` should have `a` replaced with `d`).

Assume we have this config and the equivalent working copies look like this:
```
Repo 1:
a/b/1.txt
a/2.txt

Repo 2:
c/1.txt
d/2.txt
```

Consider a bonsai changeset in repo 1, which introduces a file at path `a`. The `file_changes` section of this changeset will look like:
```
{
  MPath("a") => Some(FileChange(...))
}
```

Naive rewriting of this changeset in repo 2 will produce file changes like this:
```
{
  MPath("d") => Some(FileChange(...))
}
```

Now, if we apply both changesets to their repos' working copies:
```
Repo 1:
a

Repo 2:
c/1.txt
d
```

Common sense dictates that file `c/1.txt` in repo 2 should've been deleted, but it wasn't.

This commit addresses such problems by:
- detecting implicit deletes in the source repo
- applying `Mover`s to these deletes
- adding moved implicit deletes to the naively-rewritten file changes
- minimizing the entire set of file changes

Reviewed By: StanislavGlebik

Differential Revision: D18792591

fbshipit-source-id: 0caa6297451c58cc19d1dc82c2a42b9852ebcde4