mononoke: add --use-existing-bundle-if-available mode to hg sync job
Summary:
At the moment we have two modes - either we use a bundle that was pushed to us
by the client or sync job generates it itself. This diff adds another option
where a bundle is generated only if it's missing.

We normally won't use it, but we'd like to use it during the rollout of
megarepo when one small repo is push redirected (i.e. large repo is a source of
truth) and another repo is tailed into megarepo (i.e. small repo is a source of
truth).

Reviewed By: ikostia

Differential Revision: D18929402

fbshipit-source-id: 4594958a6efa984f49c2140f56d612482651660b