mononoke: fix flaky LFS test
Summary:
The identity parsing tests for the LFS Server have been flaky for a little
while. That's because we look for logs that are emitted asynchronously. To
mitigate the race condition, we truncate the log file, then wait for it to
become non-empty again.

This should work great in theory, but it only works kinda meh in practice if
instead of looking at the size of the log file, we look at the size of
Mononoke's `library.sh`. So, this fixes that :)

Reviewed By: HarveyHunt

Differential Revision: D18930530

fbshipit-source-id: 5823119714511298c757a82eedb57ba48fdfaba4