mononoke: fix lfs identity parsing tests harder
Summary:
Waiting for the file to be non-empty works well, but it doesn't work if the
file isn't empty because some other output was delayed and showed up after we
truncated the file.

This update our tests to wait for a specific amount of output instead to force
proper sequencing.

Reviewed By: johansglock

Differential Revision: D18973496

fbshipit-source-id: e639431daaff76e840d03f9b0f2759d2c059a74a