mononoke: make admin crossrepo map tell the user if commit is present in target
Summary:
Currently, it just informs the user if the commit is present in the mapping,
but frequently it is also interesting to know if it is directly in the target
repo (i.e. preserved).

Note: I'd love to have a subcommand to print the proper `CommitSyncOutcome`,
but that is a little more work, so for now let's keep this one.

Reviewed By: johansglock

Differential Revision: D19019273

fbshipit-source-id: 29142a6f3fdaf9e51fd187b3818d0432eac05e3b