mononoke/sql_ext: allow always going to MySQL master for reads
Summary:
A lot of our integration tests assume read-after-write consistency when talking
to Mononoke (e.g. if I update a bookmark, then pull it, then the bookmark
should be updated when I pull). This seems to result in failures when using
MySQL as the backend

To do this, I've replaced all instances of `myrouter_port` outside of sql_ext
code and replaced them with a new `MysqlOptions` struct. The idea is to make it
easier for add new flags to thread down (i.e. if someone else needs to add a
new flag, they won't have to make changes everywhere like I had to :) ).

Note that this particular diff does not use the flag anywhere just yet.

Reviewed By: StanislavGlebik

Differential Revision: D19092336

fbshipit-source-id: 47408d72e2e89218ce4183e5ecb77355aabac9bc