mononoke: Upload hg manifests concurrently during changeset generation
Summary:
hg changeset generation can be slow when there are many commits. When
generating changesets, the hg manifests are uploaded to the blobstore
sequentially.

Change this so that the writes are sent across a channel to a stream which
allows them to be buffered and uploaded concurrently. Further, add
`derive_manifest_with_blob_sender` to allow other manifest generation code to
take advantage of this speed up. A later diff will port fsnodes and unodes to
use this new function.

Using the benchmark in scm/mononoke/benchmark I've observed a 2x speed up.

Reviewed By: StanislavGlebik

Differential Revision: D18729544

fbshipit-source-id: 3119b60340bba048ffc944c2a4b83afff91f458e