mononoke: Refactor fsnodes and unodes manifest generation to use generate_manifest_with_io_sender
Summary:
Manifest generation for both fsnodes and unodes involves creating an
unbounded channel and sending blobstore operations to it in order to achieve
write concurrency. This is a common pattern that is also used for hg manifest
generation. Use `generate_manifest_with_io_sender` instead of writing the same
code for both fsnodes and unodes.

Further, add fsnodes to the generation benchmark in order to verify that this change
hasn't affected performance.

Reviewed By: krallin

Differential Revision: D18729543

fbshipit-source-id: 9866fcb8300dc309dcd8c0119e5d69b4f8915fc2