mononoke/filenodes: sort filenodes prior to inserting them
Summary:
This fixes one other instance of flakyness in the MySQL tests. The bug is that
sometimes blobimport fails with an error reporting:

```
ERROR HY000 (1205): Lock wait timeout exceeded; try restarting transaction: Timeout on record in index: $DB_NAME/paths.PRIMARY
```

The root cause of this error is a deadlock in the queries made by blobimport.
Indeed, if a given filenode needs to be inserted for two different commits,
then blobimport will dispatch queries that look like this (the number to the
left is a timestamp):

```
1576668743: WRITE_QUERY: INSERT IGNORE INTO paths (repo_id, path, path_hash) VALUES
  (0, 'b', 0xBC67B42402EED13375D1CEFA373FAE2C1077636797C1B6EA021326339943D627),
  (0, 'a', 0x6B610502D4CEF51C9506D00077FC3E1AF6BF04EFAE387773E223AF14E8FF2B29),
  (0, '', 0x78311274C58FE67167C7DD264DB43B843DC69835AC304954F0FAEEC7F30E5047);

1576668743: WRITE_QUERY: INSERT IGNORE INTO paths (repo_id, path, path_hash) VALUES
  (0, '', 0x78311274C58FE67167C7DD264DB43B843DC69835AC304954F0FAEEC7F30E5047),
  (0, 'a', 0x6B610502D4CEF51C9506D00077FC3E1AF6BF04EFAE387773E223AF14E8FF2B29);
```

As you can tell, those were dispatched at the exact same time. Right now, the
SQL adapter we use in tests establishes a new connection every time it connects
(we should improve this, but for now that's how it is), so this creates 2
connections and runs each query on one.

At a glance, this seems fine. Those queries are likely to conflict, but they
use INSERT OR IGNORE, so we expect to find ourselves with the right set of rows
inserted.

Unfortunately, you might note that the ordering in those queries is different.
As a result, what actually happens is:

- Query 1 takes the lock on the 'b' row
- Query 1 takes the lock on the 'a' row
- Query 2 takes the lock on the '' row
- Query 1 wants the lock on the '' row
  -> This is a deadlock

Surprisingly, MySQL's deadlock detection didn't kick in here.

This appears to be because it only considers deadlocks in transactions
(https://dev.mysql.com/doc/refman/8.0/en/innodb-deadlock-detection.html).

However, this particular write actually run outside of a transaction, with
`autocommit = 1`, so InnoDB doesn't see it, and we fall back to the very much
unfortunate behavior of waiting for the lock timeout (fortunately for us, our
lock timeout is pretty low at FB, but note that the default is FIFTY SECONDS!).

Anyhow. The fix here is to sort our write so that we acquire locks in order.
This causes one transaction to acquire everything it needs, and complete, and
then the other one can do its normal insert-or-ignore behavior and everything
is well and good.

Reviewed By: markbt

Differential Revision: D19158550

fbshipit-source-id: e9aca100df74001298f1fee5ebe5e5d6009a61d8