mononoke: add helper structs/function to create commits and bookmarks in tests
Summary:
It's quite hard to create bonsai changesets and move bookmarks in
unit-tests (see pushrebase tests for example). The most annoying thing was that
you quite often need to convert between HgChangesets/Bonsai changesets/bookmark
names.

This diff adds a few utils that can be used to simplify it. The best thing is
that now it's easy to create based on "master" bookmark/hg changeset/bonsai
changeset, and later we can extend to support other commit identifiers like
globalrevs.

A few notes:

1) New structs are similar to CommitIdentityScheme we use in scs_server, but there are
important differences to justify separate traits for tests:

* scs_server depends on lower-level parts of Mononoke like BlobRepo, and that
might make it impossible to use CommitIdentityScheme to test lower-level parts
of codebase because of circular dependencies.
* It doesn't support strings to refer to a commit, while it's very nice to use
it in tests.

2) These utils are supposed to be used only in tests.

3) resolve_cs_id is a copy-paste from cmdlib/helpers. Copy-paste is usually
bad, however in that case I think it's reasonable - we don't want cmdlib and
test_utils to depend on each other. Besides the function is not that big.

4) I made changes to a few pushrebase unit tests to show the improvement

Reviewed By: markbt, krallin

Differential Revision: D19175692

fbshipit-source-id: 0a6e0bc13ecb98a02f16bea16de64c43066615ff