mononoke: walker: update default walk and limit bonsai edges to the core bonsai types
Summary:
Add AliasContentMapping to the default walk so that we know they load ok and can scrub, and can add validation on them if desired.

As part of this limit the bonsai edges to the minimal bonsai set, and add new contentmeta group for the content metadata and aliases, as otherwise bonsai only walks started to include them.

Reviewed By: krallin

Differential Revision: D19091397

fbshipit-source-id: 57f5023facd2e98d19af93b47c0f48930c8fb715