common/rust/stats: refactor stats to abstract out facebook-specific code
Summary:
In order to open-source stats crate as part of rust-shed it is essential to abstract out its API to not expose Facebook internal details and code that won't be shipped to github.
This diff does that, but it does not offer yet an open-source alternative to Facebook internals - this will be added in a separate diff.

Also in order to make the macro less C++ like and more Rust like some changes were made to the types it accepts and it had to be reflected in all usages of `define_stats!`.

Reviewed By: StanislavGlebik

Differential Revision: D19158554

fbshipit-source-id: 60096341e18a0573c554dd4a681836ee04b833de