mononoke: cross repo sync - move a bookmark and insert mapping in the same txn
Summary:
Previously mononoke x repo sync job was doing two things for each commit it
needs to sync - it pushrebases a commit and then updates commit sync mapping.
This could lead to a corruption if process failed right after a successful
pushrebase.

This diff fixes it by doing the following
1) Pushrebase now allows setting sql_txn_factory. This allows
pushrebase caller create a transaction that will be used for updating a
bookmark in pushrebase.

2) Cross repo uses sql_txn_factory to create a transaction which inserts a
commit sync mapping. That means bookmark is moved only if commit sync mapping
was successfully inserted.

3) I had to update some of the tests to make sure they are created on the same
sqlite connection.

Reviewed By: krallin

Differential Revision: D19291320

fbshipit-source-id: 149d104e2fe7af6622d06abbe473e6652265b89b